Laptop/App/requirements.txt
pyserial==3.5
PySide==6.5.1.1
pyinstaller==5.12.0
pyqtgraph==0.13.3
